Done: recuperate all product from the database Done: Put all product in the list when we create a task

+++ org.eclipse.uml.to.java/src/main/java/Factory/AbstractFactory.java
@@ -46,7 +46,7 @@ public abstract class AbstractFactory {
 
 	public abstract ActivityCategory createActivityCategory(String name, String shortDetail, String longDetail ,int idUser);
 	
-	public abstract ProductJDBC createProduct(String name, float price, int quantity, String category, int idVendeur);
+	public abstract ProductJDBC createProduct(String name, Double price, int quantity, String category, int idVendeur);
 	
 	// Start of user code (user defined methods for AbstractUserFactory)
